ClickAT.java
Added some Debug Statements.
Now ClickAT will click relative to the previous position.
Click on 10,10. Now to Click on 20.20, you have to specify 10,10, since it is relative to the previous position.
Also, ClickAT is relative to where you are on the screen. If you Scroll down 1 page, the coordinates will start from the new 0,0 (which is after the scroll).